requirements.txt
celery
channels
daphne
Django
django-celery-beat
django-cors-headers
django-rest-framework
django-timezone-field
djangorestframework
djangorestframework-simplejwt
Pillow
psycopg2
PyJWT
python-dotenv
redis
mysqlclient
asgiref
pytz
sqlparse
gunicorn
dj-database-url
django-environ
uvicorn